+++ src/libprojectM/Renderer/RenderContext.hpp
@@ -6,6 +6,8 @@
 
 #include <projectM-4/projectM_cxx_export.h>
 
+#include <cstdint>
+
 namespace libprojectM {
 namespace Renderer {
 
@@ -18,13 +20,13 @@ class TextureManager;
 class PROJECTM_CXX_EXPORT RenderContext
 {
 public:
-    float time{0.0f};          //!< Time since the preset started, in seconds.
+    double time{0.0};          //!< Time since the preset started, in seconds.
     int frame{0};              //!< Frames rendered so far.
     float fps{0.0f};           //!< Frames per second.
     float progress{0.0f};      //!< Preset progress.
     float blendProgress{0.0f}; //!< Preset transition/blending progress.
-    int viewportSizeX{0};      //!< Horizontal viewport size in pixels
-    int viewportSizeY{0};      //!< Vertical viewport size in pixels
+    uint32_t viewportSizeX{0};      //!< Horizontal viewport size in pixels
+    uint32_t viewportSizeY{0};      //!< Vertical viewport size in pixels
     float aspectX{1.0};        //!< X aspect ratio.
     float aspectY{1.0};        //!< Y aspect ratio.
     float invAspectX{1.0};     //!< Inverse X aspect ratio.
